Initial work to make dynamic light max value possible
Add functions to get the maximum light duty cycle value, so that it no
longer needs to be hard-coded.

Change all light value variables from 8-bit to 16-bit.

Still need some changes to the gain calibration routines, but that will
likely be left for last.

Also need some changes for how we handle idle light settings, since
they're currently based on a set of constants.

+++ software/firmware/src/densitometer.c
@@ -83,7 +83,7 @@ void densitometer_set_idle_light(const densitometer_t *densitometer, bool enable
     if (!densitometer) { return; }
 
     if (enabled) {
-        uint8_t idle_value = 0;
+        uint16_t idle_value = 0;
 
         /* Copy over latest idle value from settings */
         sensor_light_t idle_light = densitometer->read_light;
